+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Make `device::Error` non-exhaustive.
+- Remove unused `InvalidTunnelName` and `DropPrivileges` variants of `device::Error`.
 - Move `device::Error` variants specific to the `tun` feature into `tun::tun_async_device::Error`.
 - Rename DAITA-concept of "padding packets" to "decoy packets"